Fix clientDisconnected event miss
Event firing is missed in case of a race between
ClientEngineImpl.connectionRemoved and
ClientHeartbeatMonitor.cleanupEndpointsWithDeadConnections.

Both trying to remove an endpoint(`removeEndpoint`).
If cleanupEndpointsWithDeadConnections happens to remove first,
then event is not fired.

As a fix, I moved to event firing to more centralized place so
that first call to `removeEndpoint` will fire the disconnected
event.

fixes https://github.com/hazelcast/hazelcast/issues/12674

(cherry picked from commit d6127c8c1787cecd5e87b92d1c8c1f147f7a907d)
